+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 PyBuilder SemVer Git Tag Plugin [![Build Status](https://travis-ci.org/AlexeySanko/pybuilder_semver_git_tag.svg?branch=master)](https://travis-ci.org/AlexeySanko/pybuilder_semver_git_tag)
 =======================
 
+1.0.6
+-----
+- Added tag prefix with property `semver_git_tag_version_prefix`
+- For release added possibility to check that changelog file was changed since previous release tag
+
 1.0.5
 -----
-Travis deployment added
+- Travis deployment added